refactor loadProjectFiles
